+++ example/Vector/0_simple/main.cpp
@@ -327,7 +327,10 @@ int main(int argc, char* argv[])
 	 * With this function we output the particle position in VTK format. A VTK file
 	 * contain information about particle position and properties. Such file can be visualizaed
 	 * with program like paraview. In case this program run on several processor Each processor
-	 * generate a VTK file
+	 * generate a VTK file. VTK has two format one is the ASCII that is human readable but produce
+	 * bigger file the other is the binary that produce not-human readable files, but smaller
+	 * and more efficent files to read for Paraview. In order to create a Binary VTK file use the
+	 * option VTK_
 	 *
 	 * \snippet Vector/0_simple/main.cpp vtk
 	 *
@@ -344,6 +347,7 @@ int main(int argc, char* argv[])
 	//! \cond [vtk] \endcond
 
 	vd.write("particles");
+	vd.write("particles",VTK_WRITE_BINARY);
 
 	//! \cond [vtk] \endcond